+++ backend/pom.xml
@@ -85,6 +85,10 @@
             <artifactId>mybatis-plus-annotation</artifactId>
             <version>3.4.2</version>
         </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+        </dependency>
     </dependencies>
     <dependencyManagement>
         <dependencies>
