+++ backend/src/main/java/com/wyl/backend/BackendApplication.java
@@ -6,11 +6,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@MapperScan({"com.wyl.backend.classes.problem.controller",
-        "com.wyl.backend.classes.problem.sql",
-        "com.wyl.backend.classes.user.sql",
-        "com.wyl.backend.classes.user.Controller",
-})
+@MapperScan("com.wyl.backend.mapper")
 public class BackendApplication {
 
     public static void main(String[] args) {
