--- backend/src/main/java/com/wyl/backend/classes/problem/ProblemSample.java
@@ -1,32 +1,0 @@
-package com.wyl.backend.classes.problem;
-
-public class ProblemSample {
-    private String input;
-    private String output;
-    private int problemid;
-    public int getProblemid() {
-        return problemid;
-    }
-
-    public void setProblemid(int problemid) {
-        this.problemid = problemid;
-    }
-
-
-
-    public String getInput() {
-        return input;
-    }
-
-    public void setInput(String input) {
-        this.input = input;
-    }
-
-    public String getOutput() {
-        return output;
-    }
-
-    public void setOutput(String output) {
-        this.output = output;
-    }
-}
