--- backend/src/main/java/com/wyl/backend/classes/user/userinfo/UserInfo.java
@@ -1,60 +1,0 @@
-package com.wyl.backend.classes.user.userinfo;
-
-public class UserInfo {
-    private int userid;
-    private String username;
-    private String password;
-    private String registertime;
-    private String useremail;
-    private String token;
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getUseremail() {
-        return useremail;
-    }
-
-    public void setUseremail(String useremail) {
-        this.useremail = useremail;
-    }
-
-
-
-    public int getUserid() {
-        return userid;
-    }
-
-    public void setUserid(int userid) {
-        this.userid = userid;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String userpassword) {
-        this.password = userpassword;
-    }
-
-    public String getRegistertime() {
-        return registertime;
-    }
-
-    public void setRegistertime(String registertime) {
-        this.registertime = registertime;
-    }
-}
