+++ backend/src/main/java/com/wyl/backend/entity/problem/ProblemSample.java
@@ -1,0 +1,10 @@
+package com.wyl.backend.entity.problem;
+
+import lombok.Data;
+
+@Data
+public class ProblemSample {
+    private String input;
+    private String output;
+    private int problemid;
+}
