+++ backend/src/main/java/com/wyl/backend/mapper/CreateProblemContent.java
@@ -1,8 +1,7 @@
-package com.wyl.backend.classes.problem.sql;
+package com.wyl.backend.mapper;
 
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.wyl.backend.classes.problem.ProblemContent;
-import com.wyl.backend.classes.problem.ProblemSample;
+import com.wyl.backend.entity.problem.ProblemContent;
+import com.wyl.backend.entity.problem.ProblemSample;
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
@@ -11,17 +10,21 @@ import org.apache.ibatis.annotations.Select;
 import java.util.List;
 
 @Mapper
-public interface CreateProblemContent{
+public interface CreateProblemContent {
     @Insert("INSERT INTO problemcontent(title, background, description, inputformat, outputformat, hint) " +
             "VALUES(#{problemContent.title}, #{problemContent.background}, #{problemContent.description}, " +
             "#{problemContent.inputformat}, #{problemContent.outputformat}, #{problemContent.hint})")
     int insert(@Param("problemContent") ProblemContent problemContent);
+
     @Select("SELECT * FROM problemcontent")
     List<ProblemContent> selectAll();
+
     @Select("SELECT * FROM problemcontent where problemid=#{problemid}")
     ProblemContent selectById(int problemId);
+
     @Select("SELECT * FROM problemsample where problemid=#{problemid}")
     List<ProblemSample> selectSampleById(int problemId);
+
     @Insert("INSERT INTO problemsample(problemid, input, output) " +
             "VALUES(#{problemSample.problemid}, #{problemSample.input}, #{problemSample.output})")
     int insertToSample(@Param("problemSample") ProblemSample problemSample);
