+++ backend/src/main/java/com/wyl/backend/mapper/UserOperator.java
@@ -1,7 +1,6 @@
-package com.wyl.backend.classes.user.sql;
+package com.wyl.backend.mapper;
 
-import com.wyl.backend.classes.problem.ProblemContent;
-import com.wyl.backend.classes.user.userinfo.UserInfo;
+import com.wyl.backend.entity.user.UserInfo;
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
@@ -14,6 +13,7 @@ public interface UserOperator {
     @Insert("INSERT INTO userinfo(registertime, username, password,useremail) " +
             "VALUES(#{userinfo.registertime}, #{userinfo.username}, #{userinfo.password},#{userinfo.useremail})")
     int insert(@Param("userinfo") UserInfo userinfo);
+
     @Select("SELECT * FROM userinfo")
     List<UserInfo> select();
 }
